# api/openapi/link.yml
openapi: "3.0.4"
info:
  version: 0.1.0
  title: Link API
  description: Link API Definition
  contact:
    name: Kevin Yobeth
    url: https://github.com/kevinyobeth
servers:
  - url: https://{hostname}/api
    variables:
      hostname:
        default: localhost

paths:
  /links:
    post:
      tags: [Link]
      operationId: createLink
      description: create link
      requestBody:
        description: create link
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "200":
          description: create link success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    get:
      tags: [Link]
      operationId: getLinks
      description: get links
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        "200":
          description: get links success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetLinksResponse"
  /links/{id}:
    get:
      tags: [Link]
      operationId: getLink
      description: get link
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: link id
      responses:
        "200":
          description: get link success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetLinkResponse"
    put:
      tags: [Link]
      operationId: updateLink
      description: update link
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: link slug
      requestBody:
        description: update link
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkRequest"
      responses:
        "200":
          description: update link success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      tags: [Link]
      operationId: deleteLink
      description: delete link
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
          description: link slug
      responses:
        "200":
          description: delete link success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

components:
  schemas:
    CreateLinkRequest:
      type: object
      required: [slug, url, description]
      properties:
        slug:
          type: string
        url:
          type: string
        description:
          type: string
    UpdateLinkRequest:
      type: object
      required: [slug, url, description]
      properties:
        slug:
          type: string
        url:
          type: string
        description:
          type: string

    GetLinksResponse:
      type: object
      required: [data, metadata, message]
      properties:
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Link"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
    GetLinkResponse:
      type: object
      required: [data, message]
      properties:
        message:
          type: string
        data:
          $ref: "#/components/schemas/Link"
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          
    Link:
      type: object
      required: [id, slug, url, description, total, created_at]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        url:
          type: string
        description:
          type: string
        total:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    
    PaginationMetadata:
      type: object
      properties:
        next:
          type: string
          format: uuid
        prev:
          type: string
          format: uuid
        total:
          type: integer
          format: int

  parameters:
    limit:
      name: limit
      in: query
      description: Maximum number of results to return
      required: true
      schema:
        type: integer
        format: int
        default: 10
    page:
      name: page
      in: query
      description: The page number to retrieve
      required: true
      schema:
        type: integer
        format: int
        default: 1
    next:
      name: next
      in: query
      description: The next cursor for pagination
      required: false
      schema:
        type: string
        format: uuid
    prev:
      name: prev
      in: query
      description: The previous cursor for pagination
      required: false
      schema:
        type: string
        format: uuid