# internal/link/presentation/http/openapi_types.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package http

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreateLinkRequest defines model for CreateLinkRequest.
type CreateLinkRequest struct {
	Description string `json:"description"`
	Slug        string `json:"slug"`
	Url         string `json:"url"`
}

// GetLinkResponse defines model for GetLinkResponse.
type GetLinkResponse struct {
	Data    Link   `json:"data"`
	Message string `json:"message"`
}

// GetLinksResponse defines model for GetLinksResponse.
type GetLinksResponse struct {
	Data     []Link             `json:"data"`
	Message  string             `json:"message"`
	Metadata PaginationMetadata `json:"metadata"`
}

// Link defines model for Link.
type Link struct {
	CreatedAt   time.Time          `json:"created_at"`
	Description string             `json:"description"`
	Id          openapi_types.UUID `json:"id"`
	Slug        string             `json:"slug"`
	Total       int                `json:"total"`
	UpdatedAt   *time.Time         `json:"updated_at,omitempty"`
	Url         string             `json:"url"`
}

// MessageResponse defines model for MessageResponse.
type MessageResponse struct {
	Message string `json:"message"`
}

// PaginationMetadata defines model for PaginationMetadata.
type PaginationMetadata struct {
	Next  *openapi_types.UUID `json:"next,omitempty"`
	Prev  *openapi_types.UUID `json:"prev,omitempty"`
	Total *int                `json:"total,omitempty"`
}

// UpdateLinkRequest defines model for UpdateLinkRequest.
type UpdateLinkRequest struct {
	Description string `json:"description"`
	Slug        string `json:"slug"`
	Url         string `json:"url"`
}

// Limit defines model for limit.
type Limit = int

// Page defines model for page.
type Page = int

// GetLinksParams defines parameters for GetLinks.
type GetLinksParams struct {
	// Page The page number to retrieve
	Page Page `form:"page" json:"page"`

	// Limit Maximum number of results to return
	Limit Limit `form:"limit" json:"limit"`
}

// CreateLinkJSONRequestBody defines body for CreateLink for application/json ContentType.
type CreateLinkJSONRequestBody = CreateLinkRequest

// UpdateLinkJSONRequestBody defines body for UpdateLink for application/json ContentType.
type UpdateLinkJSONRequestBody = UpdateLinkRequest